第六次提交

+++ src/main/java/com/fc/service/impl/DormitoryServiceImpl.java
@@ -2,10 +2,8 @@ package com.fc.service.impl;
 
 import com.fc.dao.BuildingMapper;
 import com.fc.dao.DormitoryMapper;
-import com.fc.entity.Building;
-import com.fc.entity.Dormitory;
-import com.fc.entity.DormitoryExample;
-import com.fc.entity.DormitoryManager;
+import com.fc.dao.LiveMapper;
+import com.fc.entity.*;
 import com.fc.service.DormitoryService;
 import com.fc.util.RandomNum;
 import com.fc.vo.DormitoryVO;
@@ -19,6 +17,9 @@ import java.util.List;
 public class DormitoryServiceImpl implements DormitoryService {
     @Autowired
     private DormitoryMapper dormitoryMapper;
+
+    @Autowired
+    private LiveMapper liveMapper;
 
     @Override
     public List findAll() {
@@ -45,8 +46,21 @@ public class DormitoryServiceImpl implements DormitoryService {
 
 
     @Override
-    public void delete(String id) {
-        dormitoryMapper.deleteByPrimaryKey(id);
+    public boolean delete(String id) {
+        LiveExample liveExample = new LiveExample();
+
+        LiveExample.Criteria criteria = liveExample.createCriteria();
+
+        criteria.andDormitoryIdEqualTo(id);
+
+        List<Live> lives = liveMapper.selectByExample(liveExample);
+
+        if (lives.size() == 0 || lives == null) {
+            dormitoryMapper.deleteByPrimaryKey(id);
+            return true;
+        }
+
+        return false;
     }
 
     @Override
